Organized the project source code. Separated the requests functions and moved them into the auth, notes, and page_render directories. Moved the LoadTemplates function into the template_loader directory.
